requirements.txt
faust-streaming==0.11.3
faust-streaming-rocksdb==0.9.3
confluent_kafka==2.10.1